refactor(gateway): extract compression-status helpers from run.py (slice 4 of #54962)
Fourth slice of the gateway god-file unpacking: extract
_status_template_to_regex and the _COMPRESSION_PROGRESS_STATUS_RE matcher
into gateway/status_helpers.py.

- Byte-identical extraction (AST-verified): the regex is built from the
  SAME template constants the emit sites format
  (agent/conversation_compression.py), so wording drift cannot silently
  diverge
- gateway/run.py: -47 lines; the 8 now-unused template imports removed;
  helpers imported at the extraction point
- 78 tests pass (compression-progress notices + replay + resume suites)

Follow-up to #77433 (slice 1) and #77438 (slices 2-3). Progress on #54962.
